Add conditional compilation for RP2040/RP2350
