+++ rpi-pico-lldb/start-tutorial/Sources/RP2350/svd2swift.json
@@ -1,0 +1,12 @@
+{
+  "peripherals": [
+    "PADS_BANK0",
+    "PIO0",
+    "IO_BANK0",
+    "RESETS",
+    "I2C0",
+    "I2C1",
+    "SIO"
+  ],
+  "access-level": "public"
+}
