--- rpi-pico2-lldb/start-tutorial/Package.swift
@@ -1,27 +1,0 @@
-// swift-tools-version: 6.0
-
-import PackageDescription
-
-let package = Package(
-  name: "rpi-pico2-neopixel",
-  products: [
-    .executable(name: "Application", targets: ["Application"])
-  ],
-  dependencies: [
-    .package(url: "https://github.com/apple/swift-mmio.git", branch: "main")
-  ],
-  targets: [
-    .executableTarget(
-      name: "Application",
-      dependencies: ["RP2350", "Support"]),
-    .target(
-      name: "RP2350",
-      dependencies: [
-        .product(name: "MMIO", package: "swift-mmio")
-      ],
-      plugins: [
-        .plugin(name: "SVD2SwiftPlugin", package: "swift-mmio")
-      ]),
-    .target(name: "Support"),
-  ],
-  swiftLanguageModes: [.v5])
